OSInfo response
