+++ src/main/kotlin/online/ruin_of_future/informative_mc_core/web_api/ApiServer.kt
@@ -18,10 +18,10 @@ package online.ruin_of_future.informative_mc_core.web_api
 import io.javalin.Javalin
 import io.javalin.apibuilder.ApiBuilder.*
 import kotlinx.serialization.ExperimentalSerializationApi
+import online.ruin_of_future.informative_mc_core.auth.TokenManager
 import online.ruin_of_future.informative_mc_core.config.ModConfig
 import online.ruin_of_future.informative_mc_core.data.ModData
 import online.ruin_of_future.informative_mc_core.tmpDirPath
-import online.ruin_of_future.informative_mc_core.auth.TokenManager
 import online.ruin_of_future.informative_mc_core.util.generateCertificate
 import online.ruin_of_future.informative_mc_core.util.generateKeyPair
 import online.ruin_of_future.informative_mc_core.web_api.handler.*
