+++ src/main/kotlin/online/ruin_of_future/informative_mc_core/web_api/handler/OSInfo.kt
@@ -15,57 +15,13 @@
  */
 package online.ruin_of_future.informative_mc_core.web_api.handler
 
-import kotlinx.serialization.Serializable
-import online.ruin_of_future.informative_mc_core.data.ModData
 import online.ruin_of_future.informative_mc_core.auth.TokenManager
-import online.ruin_of_future.informative_mc_core.util.humanReadableSize
+import online.ruin_of_future.informative_mc_core.data.ModData
 import online.ruin_of_future.informative_mc_core.web_api.ApiID
+import online.ruin_of_future.informative_mc_core.web_api.response.OSInfoResponse
 import java.io.OutputStream
 
 val OSInfoApiId = ApiID("system-info", "os-info")
-
-// TODO: Lift `requestStatus` and `requestInfo` out.
-
-@Suppress("UnUsed")
-@Serializable
-class OSInfoResponse(
-    val requestStatus: String,
-    val requestInfo: String,
-    // OS Info
-    val osName: String,
-    val maxMemory: String,
-    val allocatedMemory: String,
-    val freeMemory: String,
-) {
-    companion object {
-        fun getForNow(): OSInfoResponse {
-            return OSInfoResponse(
-                requestStatus = "success",
-                requestInfo = "",
-                osName = System.getProperty("os.name") ?: "unknown",
-                maxMemory = Runtime.getRuntime().maxMemory().humanReadableSize(),
-                allocatedMemory = Runtime.getRuntime().totalMemory().humanReadableSize(),
-                freeMemory = Runtime.getRuntime().freeMemory().humanReadableSize(),
-            )
-        }
-
-        fun unknownUser(userName: String): OSInfoResponse {
-            return OSInfoResponse(
-                requestStatus = "error",
-                requestInfo = "unknown user: $userName",
-                osName = "", maxMemory = "", allocatedMemory = "", freeMemory = "",
-            )
-        }
-
-        fun invalidToken(): OSInfoResponse {
-            return OSInfoResponse(
-                requestStatus = "error",
-                requestInfo = "invalid token",
-                osName = "", maxMemory = "", allocatedMemory = "", freeMemory = "",
-            )
-        }
-    }
-}
 
 class OSInfoHandler(
     private val tokenManager: TokenManager,
@@ -78,11 +34,11 @@ class OSInfoHandler(
     ) {
         val req = parseUserRequest(formParams)
         if (!modData.hasUserName(req.userName)) {
-            OSInfoResponse.unknownUser(req.userName).writeToStream(outputStream)
+            OSInfoResponse.unknownUserError(req.userName).writeToStream(outputStream)
         } else if (!tokenManager.verify(req.token)) {
-            OSInfoResponse.invalidToken().writeToStream(outputStream)
+            OSInfoResponse.invalidTokenError().writeToStream(outputStream)
         } else {
-            OSInfoResponse.getForNow().writeToStream(outputStream)
+            OSInfoResponse.getCurrent().writeToStream(outputStream)
         }
     }
 }
